requirements.txt
Flask
psycopg[binary]
python-dotenv
Flask-Bcrypt